requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
gunicorn==21.2.0
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==10.4.0
cloudscraper==1.2.71
google-generativeai==0.8.3
google-auth==2.37.0
